functions for preprocessing pipelien
processing pipeline for frontend (to take one screeenplay at a time)
may not be possible to finish, if so we'll go with Syrah's work

stopwords bank for preprocessing

hopefully now defunct

hopefully now defunct

currently trying to do this: LSTM (RNN) model with
preprocessed data

irrelevant

all preprocessing wrapped in one notebook

started on this but may not have time.
moving on to using sequence tokenization with tf.